use double precision

+++ src/vec2.h
@@ -1,9 +1,9 @@
 struct Vec2 {
-    float x;
-    float y;
+    double x;
+    double y;
 
     Vec2() : x(0), y(0) {}
-    Vec2(float x, float y) : x(x), y(y) {}
+    Vec2(double x, double y) : x(x), y(y) {}
 
     Vec2 operator+(const Vec2& other) const {
         return Vec2(x + other.x, y + other.y);
@@ -13,11 +13,11 @@ struct Vec2 {
         return Vec2(x - other.x, y - other.y);
     }
 
-    Vec2 operator*(float scalar) const {
+    Vec2 operator*(double scalar) const {
         return Vec2(x * scalar, y * scalar);
     }
 
-    Vec2 operator/(float scalar) const {
+    Vec2 operator/(double scalar) const {
         return Vec2(x / scalar, y / scalar);
     }
